officer set up question (backend)
