+++ lib/officer/details.dart
